bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
httpx==0.27.2
